refactor: division of data provider model
Division of the repository into two models, one responsible for obtaining raw data from the Internet and the other for translating the raw data into a project model.
