+++ lib/repositories/yelp_repository.dart
@@ -1,108 +1,17 @@
-import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
 import 'package:restaurantour/models/restaurant.dart';
-
-const _apiKey = '<PUT YOUR API KEY HERE>';
+import 'yelp_data_provider.dart';
 
 class YelpRepository {
-  late Dio dio;
-
   YelpRepository({
-    @visibleForTesting Dio? dio,
-  }) : dio = dio ??
-            Dio(
-              BaseOptions(
-                baseUrl: 'https://api.yelp.com',
-                headers: {
-                  'Authorization': 'Bearer $_apiKey',
-                  'Content-Type': 'application/graphql'
-                },
-              ),
-            );
+    @visibleForTesting YelpDataProvider? yelpDataProvider,
+  }) : yelpDataProvider = yelpDataProvider ?? YelpDataProvider();
 
-  /// Returns a response in this shape
-  /// {
-  /// "data": {
-  ///   "search": {
-  ///     "total": 5056,
-  ///     "business": [
-  ///       {
-  ///         "id": "faPVqws-x-5k2CQKDNtHxw",
-  ///         "name": "Yardbird Southern Table & Bar",
-  ///         "price": "$$",
-  ///         "rating": 4.5,
-  ///         "photos": [
-  ///           "https:///s3-media4.fl.yelpcdn.com/bphoto/_zXRdYX4r1OBfF86xKMbDw/o.jpg"
-  ///         ],
-  ///         "reviews": [
-  ///           {
-  ///             "id": "sjZoO8wcK1NeGJFDk5i82Q",
-  ///             "rating": 5,
-  ///             "user": {
-  ///               "id": "BuBCkWFNT_O2dbSnBZvpoQ",
-  ///               "image_url": "https:///s3-media2.fl.yelpcdn.com/photo/v8tbTjYaFvkzh1d7iE-pcQ/o.jpg",
-  ///               "name": "Gina T."
-  ///             }
-  ///           },
-  ///           {
-  ///             "id": "okpO9hfpxQXssbTZTKq9hA",
-  ///             "rating": 5,
-  ///             "user": {
-  ///               "id": "0x9xu_b0Ct_6hG6jaxpztw",
-  ///               "image_url": "https:///s3-media3.fl.yelpcdn.com/photo/gjz8X6tqE3e4praK4HfCiA/o.jpg",
-  ///               "name": "Crystal L."
-  ///             }
-  ///           },
-  ///        ...
-  ///     ]
-  ///   }
-  /// }
-  ///
+  final YelpDataProvider yelpDataProvider;
+
   Future<List<Restaurant>?> getRestaurants({int offset = 0}) async {
-    //
-    final response = await dio.post<Map<String, dynamic>>(
-      '/v3/graphql',
-      data: _getQuery(offset),
-    );
-    var queryResult =
-        RestaurantQueryResult.fromJson(response.data!['data']['search']);
-    return queryResult.restaurants;
-  }
-
-  String _getQuery(int offset) {
-    return '''
-query getRestaurants {
-  search(location: "Las Vegas", limit: 20, offset: $offset) {
-    total    
-    business {
-      id
-      name
-      price
-      rating
-      photos
-      reviews {
-        id
-        text
-        rating
-        user {
-          id
-          image_url
-          name
-        }
-      }
-      categories {
-        title
-        alias
-      }
-      hours {
-        is_open_now
-      }
-      location {
-        formatted_address
-      }
-    }
-  }
-}
-''';
+    var restaurantQueryResult =
+        await yelpDataProvider.getRestaurants(offset: offset);
+    return restaurantQueryResult?.restaurants;
   }
 }
